breaking review page into minor components.
