+++ frontend/src/actions/review.js
@@ -1,4 +1,4 @@
-import { FINISH_SIGNATURE_REVIEW, CLEAR_SIGNATURE_ASSETS } from '../constants/review';
+import { FINISH_SIGNATURE_REVIEW, CLEAR_SIGNATURE_ASSETS, REDO_SIGNATURE } from '../constants/review';
 
 export const finishSignatureReviewAction = () => ({
     type: FINISH_SIGNATURE_REVIEW,
@@ -6,4 +6,8 @@ export const finishSignatureReviewAction = () => ({
 
 export const clearSignatureAssetsAction = (payload) => ({
     type: CLEAR_SIGNATURE_ASSETS
+});
+
+export const redoSignatureAction = (payload) => ({
+    type: REDO_SIGNATURE
 });
