+++ frontend/src/constants/review.js
@@ -1,2 +1,3 @@
 export const FINISH_SIGNATURE_REVIEW = 'FINISH_SIGNATURE_REVIEW';
 export const CLEAR_SIGNATURE_ASSETS = 'CLEAR_SIGNATURE_ASSETS';
+export const REDO_SIGNATURE = 'REDO_SIGNATURE';
