+++ frontend/src/routes.js
@@ -5,7 +5,7 @@ import Uploadimage from './components/UploadImage/UploadImage';
 import LoginPage from './components/LoginPage';
 import ContractsPage from './components/ContractsPage';
 import RecordPage from './components/RecordPage';
-import ReviewVideoPage from './components/reviewUserVideo/';
+import ReviewVideoPage from './components/ReviewPage';
 import ScriptPage from './components/ScriptPage/Script.js'
 import React, { Component } from 'react';
 import Introduction from './components/Instructions/Introduction';
@@ -21,9 +21,9 @@ export class routes extends Component {
                 <Switch>
                     <Route path="/" exact component={LandingPage} />
                     <Route path="/contracts" component={ContractsPage} />
-                    <Route exact location={this.props.location} path='/record' component={RecordPage} />
+                    <Route path="/record" component={RecordPage} />
                     <Route path="/login" exact component={LoginPage} />
-                    <Route exact location={this.props.location} path='/review' component={ReviewVideoPage} />
+                    <Route path="/review" component={ReviewVideoPage} />
                     <Route path='/introduction' location={this.props.location} component={Introduction} />
                     <Route path='/upload_document' location={this.props.location} component={Uploadimage} />
                     <Route path="/script" component={ScriptPage} />
